Crate finchers_session

Session support for Finchers.

Supported backends:

Cookie

In-memory database

Redis (requires the feature flag feature = "redis")

Feature Flags

redis - enable Redis backend (default: off)

secure - enable signing and encryption support for Cookie values (default: on. it adds the crate ring to dependencies).

Modules

cookie
The session backend using Cookie as a session storage.

in_memory
The session backend using in-memory database.

redis
The session backend using Redis.

Structs

Session
A struct which manages the session value per request.

Traits

RawSession
The trait representing the backend to manage session value.