Struct finchers_session::cookie::CookieBackend

pub struct CookieBackend { /* fields omitted */ }

Methods

impl CookieBackend

pub fn plain() -> CookieBackend

Creates a CookieSessionBackend which stores the Cookie values as a raw form.

pub fn signed(key: Key) -> CookieBackend

Creates a CookieSessionBackend which signs the Cookie values with the specified secret key.

This method is only available if the feature flag secure is set.

pub fn private(key: Key) -> CookieBackend

Creates a CookieSessionBackend which encrypts the Cookie values with the specified secret key.

This method is only available if the feature flag secure is set.

pub fn path(
self,
value: impl Into<Cow<'static, str>>
) -> CookieBackend

Sets the path of Cookie entry.

The default value is "/".

pub fn secure(self, value: bool) -> CookieBackend

Sets the value of secure in Cookie entry.

The default value is true.

pub fn http_only(self, value: bool) -> CookieBackend

Sets the value of http_only in Cookie entry.

The default value is true.

pub fn domain(
self,
value: impl Into<Cow<'static, str>>
) -> CookieBackend

Sets the value of domain in Cookie entry.

The default value is None.

pub fn same_site(self, value: SameSite) -> CookieBackend

Sets the value of same_site in Cookie entry.

The default value is None.

pub fn max_age(self, value: Duration) -> CookieBackend

Sets the value of max_age in Cookie entry.

The default value is None.