Struct finchers_session::redis::RedisBackend

pub struct RedisBackend { /* fields omitted */ }

The instance of SessionBackend which uses Redis.

Methods

impl RedisBackend

pub fn new(client: Client) -> RedisBackend

Create a new RedisSessionBackend from the specified Redis client.

pub fn key_prefix(
self,
prefix: impl Into<String>
) -> RedisBackend

Set the prefix string used in the key name when stores the session value to Redis.

The default value is "finchers-session"

pub fn cookie_name(
self,
name: impl Into<Cow<'static, str>>
) -> RedisBackend

Set the name of Cookie entry which stores the session id.

The default value is "session-id"

pub fn timeout(self, timeout: Duration) -> RedisBackend

Set the timeout of session value.