Struct finchers_session::Session

#[must_use = "The value must be convert into a Future to finish the session handling."]

pub struct Session<S: RawSession> { /* fields omitted */ }

A struct which manages the session value per request.

Methods

impl<S> Session<S> where
S: RawSession,

pub fn new(raw: S) -> Session<S>

pub fn get(&self) -> Option<&str>

Get the session value if available.

pub fn set(
&mut self,
value: impl Into<String>
)

Set the session value.

pub fn remove(&mut self)

Annotates to remove session value to the backend.

pub fn with<R>(
self,
f: impl FnOnce(&mut Self) -> R
) -> where
R: IntoFuture<Error = Error>,